[feat/SWM-69] refactor(포스트) : 동행글 -> board, 게시글 -> post, post-item.css -> post-card.css로 변경
