+++ src/styles/components/common/board-card.css.ts
