+++ src/styles/components/common/post-card.css.ts
